Use strategic merge by default in `kubeProvider#Update`
In previous commits (as well as in comments in `kubeProvider#Update` we
have written in some detail the semantics of our Kubernetes provider's
update logic.

For details, please consult those comments. In short: when a user
submits a new version of a Kubernetes resource, `apply` must try to only
change the values the user intended to change -- that is, it must not
delete the computed values auto-populated by the API server, or values
that are injected by (e.g.) admission controllers out-of-band, before
the resource reaches the server.

For this reason, Kubernetes implements the so-called "strategic" merge,
and that is the default for Kubectl.

Until now, we have only implemented update in terms of JSON merge patch.
This means that we are not compatible with admission controllers that
update manifests out-of-band.

This commit will switch to using strategic merge by default.

Fixes #58.