Use package.json version when publishing.
Our publish scripts called `scripts/get-version`, which adds a -dirty
tag when the work-tree is dirty.  It happens that due to our source
generation stragegy, during a build in the CI the tree can become
dirty (since the generated package.json in the pack/nodejs folder gets
updated with the version of the package we are building), which means
calling `scripts/get-version` later is going to do the wrong thing.

Now, you can make an argument that simply building should not dirty
the tree, and I expect at some point that will be the case (as we
adopt Luke's versioning plan of not auto-generating version numbers),
but regardless, we really we should just be using whatever version is
in the package.json we are going to publish (since `pulumi` will use
that version when trying to download the plugins anyway).

Fixes #64