Use `pulumi.Input` in arguments to Kubernetes objects
Fixes #65. `pulumi.Input` is fundamental to the notion of composition in
Pulumi applications, allowing users to implicitly declare dependencies
between two resources (e.g., indicating that one resource must be
deployed before another).

This commit will transition Kubernetes CustomResource implementations to
use `pulumi.Input` for all members. We previously committed and backed
this change out, as it was not clear what the impact was on higher-level
libraries. In this PR chain, we believe we have resolved this issue by
adding the `getInputs` member to the CustomResource implementations.