Add `getInputs` to Kubernetes CustomResources
In Kubernetes, it's often useful for resources to share data _without_
indicating a dependency. For example, a Service might have the same
labels as the Deployment it targets; but in Kubernetes, it is
unnecessary for the Service to depend on the Deployment. Ideally we'd
like to share the data between the inputs, rather than to use `.apply`
to share the computed outputs. In this way we allow higher-level
libraries to share data between resources, but without declaring a
dependency.

This commit will hence introduce a member, `getInputs`, which will
return the raw JSON data provided as input to the CustomResource
constructor.