Use fully-qualified group names in URNs
This commit will change URNs for Kubernetes resources to use the
fully-qualified names instead of the "shorthand" name we're
(accidentally) using now.

Using the shorthand group has caused problems with (e.g.) `pulumi
delete`, since the provider needs to generate a client for a resource
with the `group` given in the URN. Since the URN contains the
"shorthand" rather than the canonical fully-qualified group name, this
process fails and the delete will terminate with an error.