Special case group name of `apiextensions`
Fixes #82. The Kubernetes group `apiextensions` should have a
`x-kubernetes-group-version-kind` member that precisely specifies the
fully-qualified group name. It does not, which is probably a bug; the
impact on our users is that they can't deploy Kubernetes CRDs.

This commit will special case this group name so that we generate Pulumi
URNs correctly.