Centralize resource await logic
Following up from PR 85, this commit will pull the Kubernetes resource
await logic into a single, centralized place. Currently this await logic
is specified in 3 switch statements that each exhaustively list every
Kubernetes group/version/kind and the await logic for each.

This commit wil also fix #87. The current behavior is to error out if
the GVK is not recognized as having specific await logic, and our
resolution for this is to simply assume the resource initialized
correctly.