Allow for replacement of resources in `core` group
Fix #68. Fix #99. Fix #81.

This commit will fix a bug that caused us to never claim that objects in
the `core` group required replacement. The root cause of this issue is
that we were using the GroupVersionKind (GVK) information from the
resource definition itself, which does not use the canonical name for
the `core` group. Instead, it uses "", the empty string.

This commit corrects this error and thereby resolves these three issues.