Implement resource autonaming
Fixes #10.

Every resource tracked by Pulumi has a unique name. When it constitutes
these resources on a Kubernetes cluster, typically these resources are
given the same name as they are given by Pulumi.

This makes many things harder. For example, if a resource must be
replaced, and Pulumi can't control the name, then that resource must be
deleted before it can be replaced. For some objects (e.g., Pods) this is
catastrophically bad, and for others (e.g., Deployments) this is not a
big deal.

This commit will add the ability of the Kubernetes provider to
auto-generate names for all resources. The Kubernetes provider
additionally will allow users to manually override this behavior by
providing a name to the resource, by setting the normal `.metadata.name`
property. When this is not set, Pulumi will manage the name itself.