Accept a `kubeconfig` config option.
This option is the literal text of a serialized `kubeconfig` value. If
this options is set, the config specified within is used in place of the
ambient config.

In addition to adding this option, these changes move the timing of
client creation, etc. to `Configure` rather than plugin creation.

Both of these changes are in service of instantiating a first-class
k8s provider that is configured based on the output of other resources
(e.g. an EKS cluster).

Fixes #100, #101.