Don't use exponential backoff in polling
Fix #107.

We use exponential backoff when polling the status of Kubernetes
resources whose initialization we're awaiting. The chief impact of this
is deleterious -- it means that a resource can successfully be
initialized, and then the polling logic could still spend ~30 seconds
before it polls again.

Since we anticipate Pulumi should normally be almost completely alone in
its direct use of the API server, and since Terraform doesn't implement
exponential backoff, this commit will remove this "feature."