Correctly generate simple dictionary types
Fix #98.

OpenAPI provides the ability to specify well-typed dictionaries, but our
codegen ignores this information. This means that things like
`.metadata.labels` are simply recorded as `object`, which can lead to
errors.

This commit uses this information to generate simple dictionary types,
i.e., types where the keys and the values both have the same type.