On cancellation, report all errors accumulated
Fixes #110. When we cancel, if there are outstanding errors (such as
failure to mount a volume, etc.), we should report those to the service
so that they can be persisted and (e.g.) reported to the user if they
try to run a `pulumi update` later without changing that resource.