Refine Deployment await logic
This commit will fix two small bugs in the Deployment await logic.

The first of these bugs is that `.spec.replicas` is an optional field in
the Deployment resource manifest, though we currently treat it as a
required field.

The second of these bugs is that the success conditions are slightly
different between a Deployment that is being created ("generation 1) and
a Deployment that is being updated ("generation > 1). The essential
difference between the two is that an update requires a rollout to
succeed. In the Deployment API, this is captured as a condition in
`.status.conditions`, called "Progressing". In the case of a Deployment
that is being created, there is no rollout, and so "Progressing" does
not appear. Since our check expects this, it will hang forever waiting
for it to appear.

To fix this, this commit will split `rolloutComplete` into two
variables, `deploymentAvailable` and `replicaSetAvailable`.
Additionally, since we can receive the message about the Deployment and
the ReplicaSet being marked available in any order, we transition to a
strategy where any update to either objects triggers a success check.

An explicit goal of this commit is to NOT change success conditions for
any Deployment whose generation > 1.