Move live object properties to the root of resource state.
We currently next the live properties of a Kubernetes resource under
the key "live" when storing the resource's state in the checkpoint file.
This prevents output properties from resolving when using Node.JS (and
most likely Python as well), as the Pulumi JS library only looks for
property values at the root of an object's state.

The simplest, most localized fix we can make for this is to move a
Kubernetes resource's live properties into the root of the resource's
state. This commit implements that approach.

Fixes #137.