Don't dirty the working tree when building
For the package.json we generate, don't embed the version string right
away. Instead, just wack the version in the one we place in the bin
folder. This is in line with how we handle our other projects.

To lock in these fixes, add similar logic to CI to ensure the worktree
is clean after a PR runs and for a push job, before we publish.