Use `/` in fully-qualified resource names
Fixes #146.

Currently we're creating canonical IDs for resources using the `.`
character. These IDs take the form of `${namespace}.${name}`, or simply
`${name}` if namespace does not exist.

We copied this strategy from ksonnet, which in turn copied it from some
part of kubectl, but this turns out to be a critical flaw, because `.`
is an allowed character in DNS-1123-compliant names. This means it's
possible to name your resource `foo.bar`, and if it does not contain a
namespace, then we'd decide the namespace was `foo`.

The solution is to move canonical IDs for resources to use `/` to
delimit namespace, which is much better because `/` is not DNS-1123
compliant, and therefore no Kubernetes resource can have the slash in
it.