Don't change object `apiVersion` when canonicalizing
To determine the owner of a `ReplicaSet` we need to canonicalize the API
version of the `ownerReference` and the relevant Deployment object. This
is because the Deployment controller will occasionally rewrite the API
version from (e.g.) apps/v1 to (e.g.) extensions/v1beta1.

Our canonicalization simply changed the core `apiVersion` field in the
objects themselves, setting it to some canonical version. This is
potentially harmful if we canonicalize to an `apiVersion` that the
server does not support. This is particularly a problem for apps/v1,
which is fairly new.

This commit will transition us to canonicalizing without any
side-effects, which ensures that we don't transition the `apiVersion`
field to a value that doesn't make sense.