Make status-rich updates actual status updates
Currently when we display an intermediate step during the initialization
of a complex resource, we sometimes emit messages as warnings, like:
"Service does not target any live Pods."

This is confusing, both at the time the message appears, when the user
wonders whether the Service is misconfigured, and when the process
terminates, where it's printed twice, once as an error and once as a
warning.

This commit will move these error messages to be more aspirational
status messages, like: "Service attempting find Pods to direct traffic
to."

Fixes #153, #154.