Fix helm chart deployment on Windows
There were some cases where we were not clear about if we needed to
escape a path or not, leading to us generating bogus paths like
`C\:\\Users\...` on Windows, because we though we needed to escape the
path seperator.

cd8b2a2a79b30938598db709c0886540f7e7f0e3 attempted to fix this but it
missed some cases. With these changes, I was able to deploy the
`kubernetes-ts-helm-wordpress` example on Windows 10, where as
previously it would just fail with some error from CreateFile (a Win32
API).

Fixes #143