Support instances of CRDs
Kubernetes allows for the extension of its API using
CustomResourceDefinitions (CRDs). Instances of CRDs are known as custom
resources.

Currently we support CRDs but not instantiating the custom resources
themselves. This commit adds support for the latter.

There are several parts required to make this work well:

* A class, `apiextensions.CustomResource`. This class only has members
  `metadata`, `kind`, and `apiVersion`, as nothing else is known
  statically ahead of time about the resource.
* By default, stop checking the schema of resources that aren't in the
  OpenAPI spec.  This is necessary because most CRDs do not have a
  schema, so instantiating a custom resource would result in a failure
  because we can't find a schema to validate against.
* (As we did in ksonnet) allow users to opt in to failing when we can't
  find the schema for some API object, using
  `kubernetes:config:checkUnknownResources`.
* Defaulting the YAML/object parsing logic to create an
  `apiextensions.CustomResource` if the GVK can't be obtained.

IMPLICATIONS: Since CRDs are necessary for Kubernetes support, and since
we have no way of statically determining what is a CRD and what is not,
outside of a type provider-like experience, this means that errors such
as nonexistent GVK, will be deferred from preview to `up`. This is not a
great user experience, but on balance it is much better than not
supporting CRDs at all.

Fixes #176. Fixes #168.