Use live objects when checking ReplicaSet has initialized
When we check whether a `Deployment` rollout has completed, we use
`.spec.replicas` from the current `Deployment`'s input. This is ok in
nearly every case, but fails in the case where this field has been
changed out-of-band. For example, if a user runs `kubectl scale`, to
scale the number of replicas down, the old inputs will still contain the
old value, but the `ReplicaSet` will be trying to achieve a smaller
value, so Pulumi will think we don't have enough replicas to consider
the rollout successful.

The solution to this problem is to check the `.spec.replicas` of the
current live `ReplicaSet` instead of the old input `Deployment`, as this
would be the plan of record.

Fixes #207.