Report creation errors even if resource exists
The current await logic for creation fails to account for the case where
an object already exists. Specifically, when we attempt to create a
resource, but one with the same name exists, we will actually report
that as an initialization error instead of a creation error.

This can lead to baffling errors. One example: because Pulumi believes
that `pulumi up` successfully created (but did not initialize) the
resource, a subsequent call to `pulumi destroy` will cause Pulumi to
delete the resource, even though it is not managed by this app, and may
not be managed by Pulumi at all. Another, more subtle example is
captured in #136.

This commit resolves this issue by changing how we detect initialization
errors in the Kubernetes provider. Currently, we will simply check if
the object exists after `await.Creation` fails, which is not reliable,
because we have no guarantee that the object was created during the call
to `await.Creation`. In this commit, we will simply require
`await.Creation` to correctly report whether the error was an
initialization error or not.

Fixes #136.