Fix panic in `Read` when API server is unreachable
Currently, when the master is unreachable, `Read` will panic when it
tries to pass a `nil` value for the live object it's read to
`checkpointObject`, the function that packages up the live and old input
values into something that can be written to the checkpoint.

The reason for this is that `Read` expects all errors to be either a
404, or an `await.InitializationError`. If the master is unreachable,
it's neither of these.

This commit changes `Read` to simply error out if attempting to read a
value from the API server has resulted in an error that is neither of
these types.

Fixes https://github.com/pulumi/pulumi/issues/1984.