Allow use of custom resources with getResource from TypeScript
TypeScript somewhat confusingly doesn't allow the implementation signature of an overload set to be used by callers.  So we need to include a more permissive `string` overload explicitly.  Note that this does mean that mis-typing a service type will lead to reduced typing, not to a TS error.  I think that is preferrable to preventing use of these functions from TS on customer resources.

Fixes #225.