Implement `CustomResource#get` in the Kubernetes provider
This commit will allow users to write something like this:

  k8s.core.v1.Service.get("kube-dashboard", "kubernetes");

A note about the semantics. The second argument is the _unique_ ID of a
Kubernetes resource. This follows convention set by other
implementations of this API. In Kubernetes, resources are uniquely
identified with the form <namespace>/<name>. If only <name> is present,
this is automatically converted to default/<name>. Importantly, because
this ID is meant to be unique, configuration variables do not affect
this -- `kubernetes:config:namespace`, for example, does not change
where we look for the resource.

Fixes #231.