Add support for headless `Service`
The await logic for `Service` hangs forever if the user supplies a
headless `Service` that has an empty selector (i.e., targets no `Pod`s).
This commit will re-use the logic we implemented for `Service`s of type
`ExternalName` (which also target 0 `Pod`s) to fix this purpose.

Background: Kubernetes exposes a notion of a "headless" `Service`, which
is a `Service` that's not associated with any IP (cluster-internal or
external). If a `.spec.selector` is not supplied, the `Service` targets
0 `Pod`s; if it is, it should target 1 or more. This commit essentially
refines the await logic to expect to target 0 `Pod`s if the service is
of type `ExternalName` OR if the `Service` is a headless service with an
empty selector.

Fixes #248.