Allow CRDs and CRs to be created in the same Pulumi app
Currently, if a CRD is created after the CRs that depend on it, the CRs
will fail. The main reason for this is that our caching discovery client
does not refresh after a CRD is created, so when asked about a G/V/K, it
will fail because it thinks the server does not know about it.

To fix this, this commit will cause us to mark the cache as invalid
any time a CRD is created. This means that the CR, when a create
operation is retried, will update its view of the server before
creation.