Default to foreground deletion only in Kubernetes 1.6
Background deletion was broken for much of Kubernetes 1.6, and our code
accordingly manually sets deletion to "foreground" after 1.6. This
causes pain when deleting CRDs with finalizers, which will hang until
the GC has a chance to delete them and all their dependencies, which is
quite likely to be a time longer than our timeout period.

This assumption is outdated for newer verions of Kubernetes, so this
commit will default to whatever the user has set for all versions > 1.7.