Replace usages of `watcher` package in favor of the client-go `watch` API in await.go (#244)
Rather than using a custom watcher for k8s resource deletion,
set up a watch for events on the specified resource. DELETE
and ERROR events are observed to provide the same semantics.
A 5 minute timeout is set, and an error is returned if a DELETE is
not observed in that time.