Process 404 errors in `Watch` as deletion
It is possible that if a CRD is deleted after the CR obtains a client
for its G/V/K, but before `Watch` is called, that `Watch` will terminate
with a 404. This should register as a deletion: if we successfully
created that CR before, and it is tracked by Pulumi, but the G/V/K is
now unregistered, or the resource is missing, we should consider this a
deletion.