Add support for Python first-class providers
Adds support for referring to the Kubernetes provider in a first-class
manner in Python. This is accomplished using a similar method to Node -
the provider is explicitly projected as a resource and users are free to
instantiate it at-will and use it to create other Kubernetes resources.

This commit also adds a smoke test for the feature.