Fix an issue with variable casing
We accidentally emitted required variable presence checks against the
property name instead of the property's language name, which resulted in
crashes due to the reference to an unbound variable. In addition, this
commit corrects the presence check to test directly against None instead
of using the bool-ness of the variable.