Fix buggy behavior with annotations/labels on preview (#461)
The unstructured GetLabels and GetAnnotations methods expect
that all of the annotations have string values, and error silently
if this is not the case. During previews that included computed
values in either the labels or annotations, the provider was
erroneously clearing these fields due to this error. This fix
manually updates those fields to avoid resetting them.