Stub out DiffConfig and CheckConfig (#495)
This commit stubs out two new RPC endpoints expected by the engine. The
new stubs are functionally equivalent to the previous behavior where the
endpoints were not provided, except that gRPC will no longer warn when
the endpoints are used.

This commit also updates the reference to pulumi/pulumi to take the new
gRPC interface changes.