Allow `apiextensions.CustomResource` to take CRD as arg
This patch is designed to make #531 vastly less likely in native Pulumi
applications.

We currently expose a `CustomResource` class which is a low-level,
mostly-untyped API for instantiating a Kubernetes CRD. Suppose a Pulumi
application declares a CRD, and also a bunch of CRs that are instances
of that CRD. Suppose then the user removes the CRD from the program, but
not the CRs. In this case, the the CRs will be "invisibly" deleted, and
subsequent `pulumi up` calls will not notice, unless they touch those
CRs, or the user runs `refresh`.

This commit allows users to pass the `CustomResourceDefinition` object
into the CR constructor, which implicitly parents the CR to the CRD.
This makes this "invisble deletion" scenario vastly less likely.