Fix namespaceable check for diff (#554)
For deployments that rely on CRDs created by a controller,
it is not possible to tell if the related CR is namespaceable
during preview or diff. The diff case was overlooked in the
previous fix for preview. Rather than returning an error,
assume that the CR is namespaceable, and correct during
the Create if required.