Implement `kubectlReplace` Invoke call
Fixes #509.

The Kubernetes provider will, with this commit, expose an `Invoke`
function, `kubernetes:kubernetes:kubectlReplace`. This function "waits"
for the equivalent of `kubectl replace` to run, which is to say, it
returns immediately.

This is primarily useful for cases when a cluster boots up with some
number of resources, and upon creation, we need to `replace` one of
those resources.

This API is purposefully not documented to discourage use.