Remove `getInputs` from type projections
`getInputs` was originally implemented before the Pulumi engine had
parallelism, and before the Kubernetes SDKs used `pulumi.Input`. The
reason for this was to allow Kuberentes objects, which are normally
well-decoupled, to reference each other without taking a dependency on
one another, avoiding the case where there is a circular dependency
(and, hence, a deadlock in updates).

This commit removes this unnecessary field.