Enable multiple instances of YAML manifests per stack (#594)
* Add optional resourcePrefix param for ConfigFile

Allow users to disambiguate resources generated from a ConfigFile
using the optional `resourcePrefix` parameter. If set, the
generated resources will be prefixed with the provided string.

* Add optional resourcePrefix param for ConfigGroup

Allow users to disambiguate resources generated from a ConfigGroup
using the optional `resourcePrefix` parameter. If set, the
generated resources will be prefixed with the provided string.

* Test resourcePrefix for ConfigFile