Return a detailed diff from Diff.
These changes add support for detailed diffs to the Kubernetes provider.
The detailed diff is calculated by computing a JSON merge patch between
the desired and new state, recursively walking the result, and recording
the difference at each leaf property as indicated by the patch.

The particular difference that is recorded for a property depends on the
old and new values:
- If the patched value is nil, the property is recorded as deleted
- If the old value is nil, the property is recorded as added
- If the types of the old and new values differ, the property is recorded
  as updated
- If both values are maps, the maps are recursively compared on a
  per-property basis and added to the diff
- If both values are arrays, the arrays are recursively compared on a
  per-element basis and added to the diff
- If both values are primitives and the values differ, the property is
  recorded as updated
- Otherwise, no diff is recorded.

If a difference is present at a path and the path matches one of the
patterns in the database of force-new properties, the diff is amended to
indicate that the resource needs to be replaced due to the change in
this property.