Fix Deployment await logic
The Deployment await logic was incorrectly checking for the
 `extensions/v1beta1` apiVersion, which was causing the
await logic to erroneously pass for change triggered rollouts,
even if the Deployment was invalid.

The fix required two changes:

1) For some reason, the resource client was defaulting to
the `extensions/v1beta1` apiVersion, even though recent
versions were available on the cluster. Updated the client
to prefer newer apiVersions.

2) Updated await logic for the old apiVersion to use the input
object rather than a received watch Event. Regardless of the
apiVersion of the received watch Event, the relevant status
field will be missing if the Deployment was created with the
`extensions/v1beta1` API.