Fix Deployment await logic for extensions/v1beta1
The Deployment await logic was not correctly checking rollout
success for the `extensions/v1beta1` apiVersion, which was
causing the await logic to erroneously pass for change
triggered rollouts, even if the Deployment was invalid.

Added additional logic to the readiness check based on
status fields present in all apiVersions.