Fix minor live object leak in await code
This commit will fix an almost-theoretical object leak in the code that
awaits resource creation, update, or read. The issue in each case is
that the final line of each of these `await` functions calls is a call
to `Get`, which will fail if the cluster is unreachable, returning `nil`
instead of an object. This causes Pulumi to believe the object was not
successfully created.

This commit will return an old version of the live object instead.

Fixes #364.