Change Service readiness condition
Previously, the await logic for a Service checked that Endpoints
targeted at least one address before being marked ready. This
had two problems:

1. The logic didn't account for `notReadyAddresses` in the
endpoints. Even if the selected Pods were unready, the
Service would be marked ready in this case.
2. The logic would not work for Deployments, etc., with replicas
set to 0. While uncommon, this is a valid scenario, and should
not cause the Service to be unready.

The new readiness condition is that the Endpoints do not include
any `notReadyAddresses`. Given the loose coupling between
Services and selected Pods, it's not possible to tell how many
Pods we are waiting for, so we are removing this as a readiness
condition. It's possible that this could erroneously mark a Service
as ready, but this seems better than the alternative of marking
a ready Service as unready.