Fix await logic for extensions/v1beta1/Deployment
The extensions/v1beta1 version of Deployment uses a different
code path because it doesn't include all of the fields we use to
check readiness on newer apiVersions. This logic was updated
in #523, but introduced a different bug in the process. The logic
should now correctly check for readiness across all Deployment
apiVersions.