Use Pyton 3 Explicitly
Use `python3` vs `python` so we always get a Python 3 runtime. Since
CI now requires the use of Python 3, I needed to force us to get a
newer version of the glcloud SDK (the cached version did not support
Python 3). We also need to set an envvar so gcloud prefers using
Python 3 in CI.