Stop using initialApiVersion annotation (#837)
In the 1.2.0 release, we added the pulumi.com/initialApiVersion annotation,
which was used internally by the provider. This caused the side effect of
requiring updates to all existing resources.

This change instead stores that information in the state file, and no longer
creates an annotation. To avoid further disruption, the annotation is not
removed if it already exists, but will be ignored by the provider.