Unify type representations in gen. (#905)
Rather than importing all types three times--once each for the
inputs, outputs, and provider APIs--simply import the types once, and
calculate the input, output, and provider API type for each property.

These changes also track whether or not a kind is nested s.t. the
unified set of types can be used to generate the input, output, and
provider APIs.

This is in preparation for Go codegen, which needs all of this
information at once.