Return error from provider for inaccessible k8s cluster
Previously, the provider erroneously expected that the
default provider pointed to a functioning Kubernetes
cluster. This led to unexpected failures in cases where
this wasn't true, such as the user manually setting the
kubeconfig value for the stack to an invalid value. This
change explicitly checks for a valid configuration, and
returns with a descriptive error if this check fails.