Improve namespaced Kind check
Create a static lookup table for all known Kinds to determine
if resources are namespace-scoped or cluster-scoped. If the
Kind is unknown (e.g., a CustomResource), attempt to look
up this information from the k8s API server. This approach
improves the reliability of the namespaced check during
previews, where the API server is often unavailable.

Additionally, move this check from a method attached to the
DynamicClientSet to a function that accepts an optional
DynamicClientSet so that it is usable in contexts without an
available API server.